trying to get a google map to render but so far no luck

+++ db/migrate/20190420064758_add_lat_to_locations.rb
@@ -1,5 +1,5 @@
 class AddLatToLocations < ActiveRecord::Migration[5.2]
   def change
-    add_column :locations, :lat, :integer
+    add_column :locations, :lat, :float
   end
 end
